huge commit. service layer and reservations implemented

+++ src/main/java/kea/sem3/jwtdemo/dto/CarRequest.java
@@ -19,7 +19,7 @@ public class CarRequest {
     //private CarBrand brand;
     private String brand;
     private String model;
-    private int pricePrDay;
+    private double pricePrDay;
     private double bestDiscount;
 }
